+++ CHANGELOG.rst
@@ -7,8 +7,14 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 [Unreleased]
 ------------
 
+Added support for tags.
+
 Added
 ^^^^^
+
+- Added ``dtool_s3.storagebroker.delete_key()`` method
+- Added ``dtool_s3.storagebroker.get_tag_key()`` method
+- Added ``dtool_s3.storagebroker.list_tags()`` method
 
 
 Changed